pyproject.toml
[tool.poetry]
name = "hw27"
version = "0.1.0"
description = ""
authors = ["TypeAnyName <alexgox11@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
psycopg2 = "^2.9.3"
Pillow = "^9.2.0"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
